Poll pg_try_advisory_lock every 100ms
This fixes an issue where `select pg_advisory_lock(x, y)` was running
indefinitely waiting for the lock. Since all PG statements are run in a
tx, this caused the tx for the running statement to also be left running
indefinitely. The result of which: vacuum couldn't clear out dead rows.

This fixes the issue and should be almost just as fast. We are sending
10 queries to the DB per second per connection though so we might be
able to think of something better, or make the polling interval configurable.